move remove mutation to its own file
[ghstack-poisoned]
